requirements.txt
accelerate==0.24.1
aiohttp==3.8.6
aiosignal==1.3.1
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
caldav==1.3.6
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cmake==3.27.7
datasets==2.14.6
dill==0.3.7
discord==2.3.2
discord.py==2.3.2
distro==1.8.0
docker-pycreds==0.4.0
fastapi==0.104.1
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.1
httptools==0.6.1
httpx==0.25.1
huggingface-hub==0.17.3
icalendar==5.0.11
idna==3.4
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
lit==17.0.4
llvmlite==0.41.1
lxml==4.9.3
markdownify==0.11.6
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
ninja==1.11.1.1
numba==0.58.1
numpy==1.26.1
openai==1.2.2
orjson==3.9.10
packaging==23.2
pandas==2.1.2
protobuf==4.25.0
psutil==5.9.6
pyarrow==14.0.1
pydantic==1.10.13
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
ray==2.8.0
recurring-ical-events==2.1.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rpds-py==0.12.0
safetensors==0.4.0
sentencepiece==0.1.99
sentry-sdk==1.34.0
setproctitle==1.3.3
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.14.1
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
vllm==0.2.1.post1
vobject==0.9.6.1
wandb==0.16.0
watchfiles==0.21.0
websockets==12.0
x-wr-timezone==0.0.6
xformers==0.0.22
xxhash==3.4.1
yarl==1.9.2
